disabling the scale tracker for the moment
